requirements.txt
dash
dash_bootstrap_components
plotly
pandas
sqlalchemy
python-dotenv
psycopg2-binary